Proper implementation of sand simulation
